Fix platform integration tests by moving aggregation classes to separate package
