+++ molgenis-data-elasticsearch/src/main/java/org/molgenis/data/elasticsearch/SearchService.java
@@ -1,6 +1,8 @@
 package org.molgenis.data.elasticsearch;
 
 import org.molgenis.data.*;
+import org.molgenis.data.aggregation.AggregateQuery;
+import org.molgenis.data.aggregation.AggregateResult;
 import org.molgenis.data.elasticsearch.ElasticsearchService.IndexingMode;
 import org.molgenis.data.meta.model.EntityType;
