+++ molgenis-data-elasticsearch/src/test/java/org/molgenis/data/elasticsearch/IndexedRepositoryDecoratorTest.java
@@ -2,6 +2,7 @@ package org.molgenis.data.elasticsearch;
 
 import org.molgenis.data.*;
 import org.molgenis.data.QueryRule.Operator;
+import org.molgenis.data.aggregation.AggregateQuery;
 import org.molgenis.data.meta.model.Attribute;
 import org.molgenis.data.meta.model.EntityType;
 import org.molgenis.data.support.AggregateQueryImpl;
