+++ molgenis-data-mapper/src/main/java/org/molgenis/data/mapper/controller/MappingServiceController.java
@@ -4,6 +4,7 @@ import com.google.common.collect.*;
 import org.apache.commons.lang3.StringUtils;
 import org.molgenis.auth.User;
 import org.molgenis.data.*;
+import org.molgenis.data.aggregation.AggregateResult;
 import org.molgenis.data.importer.ImportWizardController;
 import org.molgenis.data.mapper.data.request.GenerateAlgorithmRequest;
 import org.molgenis.data.mapper.data.request.MappingServiceRequest;
