+++ molgenis-data-rest/src/main/java/org/molgenis/data/rest/v2/EntityAggregatesResponse.java
@@ -1,8 +1,8 @@
 package org.molgenis.data.rest.v2;
 
-import org.molgenis.data.AggregateAnonymizer;
-import org.molgenis.data.AggregateResult;
-import org.molgenis.data.AnonymizedAggregateResult;
+import org.molgenis.data.aggregation.AggregateAnonymizer;
+import org.molgenis.data.aggregation.AggregateResult;
+import org.molgenis.data.aggregation.AnonymizedAggregateResult;
 import org.molgenis.data.Entity;
 import org.molgenis.data.meta.model.Attribute;
