+++ molgenis-data-validation/src/main/java/org/molgenis/data/validation/RepositoryValidationDecorator.java
@@ -2,6 +2,8 @@ package org.molgenis.data.validation;
 
 import org.molgenis.data.*;
 import org.molgenis.data.QueryRule.Operator;
+import org.molgenis.data.aggregation.AggregateQuery;
+import org.molgenis.data.aggregation.AggregateResult;
 import org.molgenis.data.meta.model.Attribute;
 import org.molgenis.data.meta.model.EntityType;
 import org.molgenis.data.support.QueryImpl;
