+++ molgenis-data/src/main/java/org/molgenis/data/DataService.java
@@ -1,5 +1,7 @@
 package org.molgenis.data;
 
+import org.molgenis.data.aggregation.AggregateQuery;
+import org.molgenis.data.aggregation.AggregateResult;
 import org.molgenis.data.meta.MetaDataService;
 import org.molgenis.data.meta.model.EntityType;
