+++ molgenis-data/src/main/java/org/molgenis/data/RepositorySecurityDecorator.java
@@ -1,6 +1,8 @@
 package org.molgenis.data;
 
 import org.molgenis.data.QueryRule.Operator;
+import org.molgenis.data.aggregation.AggregateQuery;
+import org.molgenis.data.aggregation.AggregateResult;
 import org.molgenis.data.meta.model.EntityType;
 import org.molgenis.data.support.QueryImpl;
 import org.molgenis.security.core.Permission;
