+++ molgenis-data/src/main/java/org/molgenis/data/aggregation/AggregateAnonymizerImpl.java
@@ -1,10 +1,10 @@
-package org.molgenis.data.support;
+package org.molgenis.data.aggregation;
 
 import com.google.common.base.Function;
 import com.google.common.collect.Lists;
-import org.molgenis.data.AggregateAnonymizer;
-import org.molgenis.data.AggregateResult;
-import org.molgenis.data.AnonymizedAggregateResult;
+import org.molgenis.data.aggregation.AggregateAnonymizer;
+import org.molgenis.data.aggregation.AggregateResult;
+import org.molgenis.data.aggregation.AnonymizedAggregateResult;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
