+++ molgenis-data/src/main/java/org/molgenis/data/aggregation/AggregateQuery.java
@@ -1,5 +1,7 @@
-package org.molgenis.data;
+package org.molgenis.data.aggregation;
 
+import org.molgenis.data.Entity;
+import org.molgenis.data.Query;
 import org.molgenis.data.meta.model.Attribute;
 
 public interface AggregateQuery
