+++ molgenis-data/src/main/java/org/molgenis/data/aggregation/AggregateResult.java
@@ -1,4 +1,4 @@
-package org.molgenis.data;
+package org.molgenis.data.aggregation;
 
 import java.util.List;
