+++ molgenis-data/src/main/java/org/molgenis/data/listeners/EntityListenerRepositoryDecorator.java
@@ -2,6 +2,8 @@ package org.molgenis.data.listeners;
 
 import org.molgenis.data.*;
 import org.molgenis.data.QueryRule.Operator;
+import org.molgenis.data.aggregation.AggregateQuery;
+import org.molgenis.data.aggregation.AggregateResult;
 import org.molgenis.data.meta.model.EntityType;
 
 import java.io.IOException;
