+++ molgenis-data/src/main/java/org/molgenis/data/mem/InMemoryRepository.java
@@ -3,6 +3,8 @@ package org.molgenis.data.mem;
 import com.google.common.collect.Sets;
 import org.molgenis.data.*;
 import org.molgenis.data.QueryRule.Operator;
+import org.molgenis.data.aggregation.AggregateQuery;
+import org.molgenis.data.aggregation.AggregateResult;
 import org.molgenis.data.meta.model.EntityType;
 import org.molgenis.data.support.QueryImpl;
