+++ molgenis-data/src/main/java/org/molgenis/data/meta/AttributeRepositoryDecorator.java
@@ -2,6 +2,8 @@ package org.molgenis.data.meta;
 
 import org.molgenis.AttributeType;
 import org.molgenis.data.*;
+import org.molgenis.data.aggregation.AggregateQuery;
+import org.molgenis.data.aggregation.AggregateResult;
 import org.molgenis.data.meta.model.Attribute;
 import org.molgenis.data.meta.model.AttributeMetadata;
 import org.molgenis.data.meta.model.EntityType;
