+++ molgenis-data/src/main/java/org/molgenis/data/meta/MetaDataRepositoryDecorator.java
@@ -2,6 +2,8 @@ package org.molgenis.data.meta;
 
 import org.molgenis.data.*;
 import org.molgenis.data.QueryRule.Operator;
+import org.molgenis.data.aggregation.AggregateQuery;
+import org.molgenis.data.aggregation.AggregateResult;
 import org.molgenis.data.meta.model.EntityType;
 
 import java.io.IOException;
