+++ molgenis-data/src/main/java/org/molgenis/data/rsql/AggregateQueryRsql.java
@@ -1,7 +1,7 @@
 package org.molgenis.data.rsql;
 
 import cz.jirutka.rsql.parser.ast.Node;
-import org.molgenis.data.AggregateQuery;
+import org.molgenis.data.aggregation.AggregateQuery;
 import org.molgenis.data.Entity;
 import org.molgenis.data.Query;
 import org.molgenis.data.meta.model.EntityType;
