+++ molgenis-data/src/main/java/org/molgenis/data/support/AbstractRepository.java
@@ -4,6 +4,8 @@ import com.google.common.collect.Iterators;
 import com.google.common.collect.Sets;
 import org.molgenis.data.*;
 import org.molgenis.data.QueryRule.Operator;
+import org.molgenis.data.aggregation.AggregateQuery;
+import org.molgenis.data.aggregation.AggregateResult;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.transaction.annotation.Transactional;
