+++ molgenis-data/src/main/java/org/molgenis/data/support/AggregateQueryImpl.java
@@ -1,6 +1,6 @@
 package org.molgenis.data.support;
 
-import org.molgenis.data.AggregateQuery;
+import org.molgenis.data.aggregation.AggregateQuery;
 import org.molgenis.data.Entity;
 import org.molgenis.data.Query;
 import org.molgenis.data.meta.model.Attribute;
