+++ molgenis-data/src/main/java/org/molgenis/data/transaction/TransactionalRepositoryDecorator.java
@@ -1,6 +1,8 @@
 package org.molgenis.data.transaction;
 
 import org.molgenis.data.*;
+import org.molgenis.data.aggregation.AggregateQuery;
+import org.molgenis.data.aggregation.AggregateResult;
 import org.springframework.transaction.PlatformTransactionManager;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.support.TransactionTemplate;
