+++ molgenis-data/src/test/java/org/molgenis/data/EntityReferenceResolverDecoratorTest.java
@@ -3,6 +3,7 @@ package org.molgenis.data;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Captor;
 import org.mockito.Mock;
+import org.molgenis.data.aggregation.AggregateQuery;
 import org.molgenis.data.meta.model.EntityType;
 import org.molgenis.data.support.QueryImpl;
 import org.testng.annotations.BeforeMethod;
