+++ molgenis-data/src/test/java/org/molgenis/data/RepositorySecurityDecoratorTest.java
@@ -1,6 +1,8 @@
 package org.molgenis.data;
 
 import org.mockito.ArgumentCaptor;
+import org.molgenis.data.aggregation.AggregateQuery;
+import org.molgenis.data.aggregation.AggregateResult;
 import org.molgenis.data.meta.model.EntityType;
 import org.springframework.security.authentication.TestingAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
