+++ molgenis-data/src/test/java/org/molgenis/data/aggregation/AggregateAnonymizerRepositoryDecoratorTest.java
@@ -1,5 +1,8 @@
-package org.molgenis.data;
+package org.molgenis.data.aggregation;
 
+import org.molgenis.data.Entity;
+import org.molgenis.data.Repository;
+import org.molgenis.data.aggregation.*;
 import org.molgenis.data.settings.AppSettings;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
