+++ molgenis-file-ingester/src/main/java/org/molgenis/file/ingest/FileIngestRepositoryDecorator.java
@@ -2,6 +2,8 @@ package org.molgenis.file.ingest;
 
 import org.molgenis.data.*;
 import org.molgenis.data.QueryRule.Operator;
+import org.molgenis.data.aggregation.AggregateQuery;
+import org.molgenis.data.aggregation.AggregateResult;
 import org.molgenis.data.meta.model.EntityType;
 import org.molgenis.file.ingest.meta.FileIngestJobExecutionMetaData;
 import org.molgenis.file.ingest.meta.FileIngestMetaData;
