+++ molgenis-platform-integration-tests/src/test/java/org/molgenis/integrationtest/platform/PlatformITConfig.java
@@ -60,7 +60,8 @@ import static org.molgenis.integrationtest.platform.PostgreSqlDatabase.dropAndCr
  FIXME Ideally, we'd like to scan all of org.molgenis.data or even org.molgenis, but there's some unwanted dependencies
  in org.molgenis.data and subpackages from included modules
   */
-@ComponentScan({ "org.molgenis.data.meta", "org.molgenis.data.index", "org.molgenis.data.jobs",
+@ComponentScan({ "org.molgenis.data.aggregation", "org.molgenis.data.meta", "org.molgenis.data.index",
+		"org.molgenis.data.jobs",
 		"org.molgenis.data.elasticsearch", "org.molgenis.auth", "org.molgenis.test.data", "org.molgenis.data.platform",
 		"org.molgenis.data.meta.model", "org.molgenis.data.system.model", "org.molgenis.data.cache",
 		"org.molgenis.data.i18n", "org.molgenis.data.postgresql", "org.molgenis.file.model",
