+++ molgenis-security/src/main/java/org/molgenis/security/owned/OwnedEntityRepositoryDecorator.java
@@ -2,6 +2,8 @@ package org.molgenis.security.owned;
 
 import org.molgenis.data.*;
 import org.molgenis.data.QueryRule.Operator;
+import org.molgenis.data.aggregation.AggregateQuery;
+import org.molgenis.data.aggregation.AggregateResult;
 import org.molgenis.data.meta.model.EntityType;
 import org.molgenis.data.support.QueryImpl;
 import org.molgenis.security.core.runas.SystemSecurityToken;
